refactor to shared scripts, add support for multiple seed files
