+++ functions/.gitignore
@@ -20,6 +20,3 @@ dist
 # Emulator data
 data/emulated
 data/exported
-data/seed/*
-!data/seed/.gitkeep
-!data/seed/seed-default.zip
