+++ functions/package.json
@@ -13,9 +13,7 @@
     "watch": "npm run precompile && webpack --watch",
     "webpack": "webpack",
     "precompile": "ts-node scripts/precompile.ts",
-    "db:backup:local": "firestore-export -a service.json -b backup.json -p",
-    "emulator:seed": "ts-node scripts/emulator/seed.ts",
-    "emulator:generate:zip": "ts-node scripts/emulator/generate-zip.ts"
+    "db:backup:local": "firestore-export -a service.json -b backup.json -p"
   },
   "main": "index.js",
   "dependencies": {
@@ -30,8 +28,8 @@
     "google-auth-library": "^6.1.1",
     "googleapis": "^61.0.0",
     "log-update": "^4.0.0",
-    "oa-shared": "^1.0.0",
-    "one-army-community-platform": "^1",
+    "oa-shared": "workspace:*",
+    "one-army-community-platform": "workspace:*",
     "request": "^2.88.2",
     "sharp": "^0.26.1"
   },
@@ -47,13 +45,11 @@
     "@types/sharp": "^0.26.0",
     "@types/webpack": "^4.41.29",
     "@types/webpack-node-externals": "^2.5.1",
-    "archiver": "^5.3.0",
     "awesome-typescript-loader": "^5.2.1",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
     "concurrently": "^6.2.0",
     "cross-env": "^7.0.3",
     "generate-json-webpack-plugin": "^1.0.0",
-    "jszip": "^3.7.0",
     "ts-node": "^10.0.0",
     "tslint": "^6.1.3",
     "typescript": "4.2.2",
