+++ scripts/package.json
@@ -2,14 +2,23 @@
   "name": "oa-scripts",
   "version": "1.0.0",
   "scripts": {
-    "install:clean": "ts-node ./install-clean.ts"
+    "install:clean": "ts-node ./install-clean.ts",
+    "emulator:seed": "ts-node emulator/seed.ts",
+    "emulator:generate:zip": "ts-node emulator/generate-zip.ts"
   },
   "dependencies": {
+    "archiver": "^5.3.0",
+    "fs-extra": "^9.0.1",
     "globby": "^11.0.2",
+    "inquirer": "^7.3.3",
+    "jszip": "^3.7.0",
     "oa-shared": "workspace:*",
-    "rimraf": "^3.0.2"
+    "rimraf": "^3.0.2",
+    "run-async": "^2.4.1"
   },
   "devDependencies": {
+    "@types/fs-extra": "^9.0.2",
+    "@types/inquirer": "^8.1.3",
     "@types/rimraf": "^3.0.2",
     "ts-node": "^10.2.1"
   }
