+++ shared/tsconfig.json
@@ -2,7 +2,7 @@
   // Note - this allows the shared workspace to be built and imported into other workspaces
   "compilerOptions": {
     "outDir": "./lib",
-    "typeRoots": ["../node_modules/@types"],
+    "typeRoots": ["./node_modules/@types"],
     "composite": true,
     "skipLibCheck": true,
     "rootDir": "."
